Move program enrolment query to the programs package
