+++ client/packages/programs/src/api/hooks/utils/useProgramEnrolmentApi.ts
@@ -1,11 +1,16 @@
 import { useGql, useAuthContext } from '@openmsupply-client/common';
-import { getProgramEnrolmentQueries } from '../../api';
+import {
+  getProgramEnrolmentQueries,
+  ProgramEnrolmentListParams,
+} from '../../api';
 import { getSdk } from '../../operations.generated';
 
 export const useProgramEnrolmentApi = () => {
   const { storeId } = useAuthContext();
   const keys = {
-    base: () => ['program-enrolment'] as const,
+    base: () => ['programEnrolment'] as const,
+    list: (params: ProgramEnrolmentListParams) =>
+      [...keys.base(), storeId, 'list', params] as const,
   };
   const { client } = useGql();
   const queries = getProgramEnrolmentQueries(getSdk(client), storeId);
