+++ client/packages/programs/src/api/index.ts
@@ -1,5 +1,6 @@
 export {
   DocumentFragment,
   DocumentRegistryFragment,
+  ProgramEventFragment,
 } from './operations.generated';
 export * from './hooks';
