+++ client/packages/system/src/Patient/ProgramEnrolment/api/index.ts
@@ -1,10 +1,1 @@
-import { ProgramEnrolmentRowFragment } from '../../api/operations.generated';
-export {
-  ProgramEnrolmentRowFragment,
-  ProgramEventFragment,
-} from '../../api/operations.generated';
 export * from './hooks';
-export { ProgramEnrolmentFragment } from './operations.generated';
-export type ProgramEnrolmentRowFragmentWithId = {
-  id: string;
-} & ProgramEnrolmentRowFragment;
