+++ client/packages/system/src/Patient/api/hooks/document/index.ts
@@ -5,7 +5,6 @@ import { useInsertPatient } from './useInsertPatient';
 import { useUpdatePatient } from './useUpdatePatient';
 import { useDocumentHistory } from './useDocumentHistory';
 import { useEncounters } from './useEncounters';
-import { useProgramEnrolments } from './useProgramEnrolments';
 import { useProgramEncounters } from './useProgramEncounters';
 
 export const Document = {
@@ -16,6 +15,5 @@ export const Document = {
   useUpdatePatient,
   useDocumentHistory,
   useEncounters,
-  useProgramEnrolments,
   useProgramEncounters,
 };
