+++ client/packages/system/src/Patient/api/hooks/index.ts
@@ -16,7 +16,6 @@ export const usePatient = {
     insert: Document.useInsertPatient,
     update: Document.useUpdatePatient,
     history: Document.useDocumentHistory,
-    programEnrolments: Document.useProgramEnrolments,
     programEncounters: Document.useProgramEncounters,
   },
 };
